apps/worker/requirements.txt
azure-functions==1.21.3
httpx==0.28.1

